+++ advanced_alchemy/types/__init__.py
@@ -1,13 +1,12 @@
 """SQLAlchemy custom types for use with the ORM."""
 
-from advanced_alchemy.types import file_object
+from advanced_alchemy.types import encrypted_string, file_object, password_hash
 from advanced_alchemy.types.datetime import DateTimeUTC
 from advanced_alchemy.types.encrypted_string import (
     EncryptedString,
     EncryptedText,
     EncryptionBackend,
     FernetBackend,
-    PGCryptoBackend,
 )
 from advanced_alchemy.types.file_object import (
     FileObject,
@@ -22,6 +21,7 @@ from advanced_alchemy.types.guid import GUID, NANOID_INSTALLED, UUID_UTILS_INSTA
 from advanced_alchemy.types.identity import BigIntIdentity
 from advanced_alchemy.types.json import ORA_JSONB, JsonB
 from advanced_alchemy.types.mutables import MutableList
+from advanced_alchemy.types.password_hash.base import HashedPassword, PasswordHash
 
 __all__ = (
     "GUID",
@@ -36,13 +36,16 @@ __all__ = (
     "FernetBackend",
     "FileObject",
     "FileObjectList",
+    "HashedPassword",
     "JsonB",
     "MutableList",
-    "PGCryptoBackend",
+    "PasswordHash",
     "StorageBackend",
     "StorageBackendT",
     "StorageRegistry",
     "StoredObject",
+    "encrypted_string",
     "file_object",
+    "password_hash",
     "storages",
 )